Code clean-up: move initialization to main.c
